requirements.txt
boto3
pdf2image
imagehash
tqdm
python-dotenv
Pillow
